Extract tests for isEvalError to own file
